simple json authentication added. performer mutation tests added

+++ music_review/graphql/album/mutations.py
@@ -1,6 +1,8 @@
 import graphene
+from graphql_jwt.decorators import login_required
 
 from .types import AlbumType
+from ..utils import check_permissions
 from ...reviews.models import Album
 
 
@@ -15,9 +17,10 @@ class CreateAlbum(graphene.relay.ClientIDMutation):
         description = graphene.String(default="")
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, performer: str, **kwargs):
         performer = graphene.relay.Node.get_node_from_global_id(info, performer)
-        album = Album.objects.create(performer=performer, **kwargs,)
+        album = Album.objects.create(performer=performer, user=info.context.user, **kwargs, )
         return CreateAlbum(album=album)
 
 
@@ -33,9 +36,11 @@ class UpdateAlbum(graphene.relay.ClientIDMutation):
         description = graphene.String()
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, album: str, **kwargs):
         kwargs_cpy = kwargs.copy()
         album_instance = graphene.relay.Node.get_node_from_global_id(info, album)
+        check_permissions(album_instance.user, info)
         if performer_id := kwargs_cpy.get("performer"):
             kwargs_cpy["performer"] = graphene.relay.Node.get_node_from_global_id(
                 info, performer_id
@@ -53,8 +58,10 @@ class DeleteAlbum(graphene.relay.ClientIDMutation):
         album = graphene.ID(required=True)
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, album: str):
         album_instance = graphene.relay.Node.get_node_from_global_id(info, album)
+        check_permissions(album_instance.user, info)
         if album_instance is None:
             return DeleteAlbum(success=False)
         album_instance.delete()
