+++ music_review/graphql/performer/mutations.py
@@ -1,8 +1,10 @@
 from typing import Optional, List, Dict
 
 import graphene
+from graphql_jwt.decorators import login_required
 
 from .types import PerformerType
+from ..utils import check_permissions
 from ...reviews.models import Performer, Album
 
 
@@ -23,17 +25,18 @@ class CreatePerformer(graphene.relay.ClientIDMutation):
         albums = graphene.List(graphene.NonNull(AlbumInputType), default=[])
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(
         cls,
-        root,
+        _,
         info,
         name: str,
-        logo_url: Optional[str],
-        description: str,
-        albums: List[AlbumInputType],
+        logo_url: Optional[str] = None,
+        description: Optional[str] = "",
+        albums: List[AlbumInputType] = (),
     ):
         performer = Performer.objects.create(
-            name=name, logo_url=logo_url, description=description,
+            name=name, logo_url=logo_url, description=description, user=info.context.user
         )
         for album in albums:
             Album.objects.create(
@@ -42,6 +45,7 @@ class CreatePerformer(graphene.relay.ClientIDMutation):
                 year=album.year,
                 cover_url=album.cover_url,
                 description=album.description,
+                user=info.context.user,
             )
 
         return CreatePerformer(performer=performer)
@@ -57,10 +61,12 @@ class UpdatePerformer(graphene.relay.ClientIDMutation):
         description = graphene.String()
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, performer: str, **kwargs: Dict[str, str]):
         performer_instance = graphene.relay.Node.get_node_from_global_id(
             info, performer
         )
+        check_permissions(performer_instance.user, info)
         for key, value in kwargs.items():
             setattr(performer_instance, key, value)
         performer_instance.save()
@@ -74,10 +80,12 @@ class DeletePerformer(graphene.relay.ClientIDMutation):
         performer = graphene.ID(required=True)
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, performer: str):
         performer_instance = graphene.relay.Node.get_node_from_global_id(
             info, performer
         )
+        check_permissions(performer_instance.user, info)
         if performer_instance is None:
             return DeletePerformer(success=False)
         performer_instance.delete()
