+++ music_review/graphql/review/mutations.py
@@ -1,6 +1,8 @@
 import graphene
+from graphql_jwt.decorators import login_required
 
 from .types import ReviewType
+from ..utils import check_permissions
 from ...reviews.models import Review
 
 
@@ -13,9 +15,10 @@ class CreateReview(graphene.relay.ClientIDMutation):
         rating = graphene.Float(required=True)
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, album: str, **kwargs):
         album = graphene.relay.Node.get_node_from_global_id(info, album)
-        review = Review.objects.create(album=album, **kwargs,)
+        review = Review.objects.create(album=album, user=info.context.user, **kwargs, )
         return CreateReview(review=review)
 
 
@@ -29,9 +32,11 @@ class UpdateReview(graphene.relay.ClientIDMutation):
         rating = graphene.Float()
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, review_id: str, **kwargs):
         kwargs_cpy = kwargs.copy()
         review_instance = graphene.relay.Node.get_node_from_global_id(info, review_id)
+        check_permissions(review_instance.user, info)
         if album_id := kwargs_cpy.get("album"):
             kwargs_cpy["album"] = graphene.relay.Node.get_node_from_global_id(
                 info, album_id
@@ -49,8 +54,10 @@ class DeleteReview(graphene.relay.ClientIDMutation):
         review = graphene.ID(required=True)
 
     @classmethod
+    @login_required
     def mutate_and_get_payload(cls, _, info, review: str):
         review_instance = graphene.relay.Node.get_node_from_global_id(info, review)
+        check_permissions(review_instance.user, info)
         if review_instance is None:
             return DeleteReview(success=False)
         review_instance.delete()
